add mission test

+++ T-Manager/src/main/java/cpt202/groupwork/controller/MissionController.java
@@ -9,6 +9,7 @@ import cpt202.groupwork.entity.Gantt;
 import cpt202.groupwork.entity.Mission;
 import cpt202.groupwork.entity.Todo;
 import cpt202.groupwork.entity.Todolist;
+import cpt202.groupwork.entity.User;
 import cpt202.groupwork.repository.GanttRepository;
 import cpt202.groupwork.repository.MissionRepository;
 import cpt202.groupwork.repository.ProjectRepository;
@@ -57,18 +58,23 @@ public class MissionController {
   GanttRepository ganttRepository;
 
   @PostMapping("/add")
-  @Operation(summary = "通过 ganttId 和 missionDTO 添加 mission")
+  @Operation(summary = "add mission with ganttId and missionDTO ")
   public Response<?> createMission(@Valid @RequestBody MissionDTO missionDTO) {
-//    Optional<String> username = SecurityUtils.getCurrentUsername();
-//    if (username.isEmpty()) {
-//      return Response.unAuth();
-//    }
+
     Integer ganttId = missionDTO.getGanttId();
     Optional<Gantt> gantt = ganttRepository.findById(ganttId);
+
+    if (gantt.equals(Optional.empty())) {
+      return Response.exceptionHandling(301, "gantt does not exist");
+    }
+    int nameLength = missionDTO.getMissionName().length();
+    if (nameLength < 1 || nameLength > 20) {
+      return Response.exceptionHandling(342, "mission name should be between 1 and 20 characters");
+    }
+
     Mission mission = new Mission();
     BeanUtils.copyProperties(missionDTO, mission);
-    //mission.setMissionProgress(0);
-    //mission.setMissionDuration(1);
+
     mission.setGanttId(ganttId);
     gantt.get().setGanttTotalNum(gantt.get().getGanttTotalNum() + 1);
     ganttRepository.save(gantt.get());
@@ -77,21 +83,15 @@ public class MissionController {
   }
 
   @DeleteMapping("/delete/{missionId}")
-  @Operation(summary = "删除mission")
+  @Operation(summary = "delete mission")
   public Response<?> deleteMission(@PathVariable Long missionId) {
-//    Optional<String> username = SecurityUtils.getCurrentUsername();
-//    if (username.isEmpty()) {
-//      return Response.unAuth();
-//    }
+
     Optional<Mission> mission = missionRepository.findById(missionId);
+
+    if (mission.equals(Optional.empty())) {
+      return Response.exceptionHandling(301, "mission does not exist");
+    }
     Gantt gantt = ganttRepository.findById(mission.get().getGanttId()).get();
-//    if (todo.isEmpty()) {
-//      return Response.ok();
-//    }
-    //只有自己才能删除自己的todo的限制
-//    if (!username.get().equals(subDiscussion.get().getUsername())) {
-//      return TeaInfo.permissionDenied("只有自己才能删除哦！");
-//    }
     gantt.setGanttTotalNum(gantt.getGanttTotalNum() - 1);
     ganttRepository.save(gantt);
     missionRepository.deleteById(missionId);
@@ -99,19 +99,20 @@ public class MissionController {
   }
 
   @PutMapping("/edit")
-  @Operation(summary = "修改mission信息")
+  @Operation(summary = "modify mission information")
   public Response<?> putMission(@Valid @RequestBody Mission missionInfo) {
-//    Optional<String> username = SecurityUtils.getCurrentUsername();
-    // 没有登陆
-//    if (username.isEmpty()) {
-//      return Response.unAuth();
-//    }
+
     Long missionId = missionInfo.getMissionId();
     Optional<Mission> mission = missionRepository.findById(missionId);
 
-//    if (mission.isEmpty()) {
-//      return Response.notFound("没有找到todo哦！");
-//    }
+    if (mission.equals(Optional.empty())) {
+      return Response.exceptionHandling(301, "mission does not exist");
+    }
+    int nameLength = missionInfo.getMissionName().length();
+    if (nameLength < 1 || nameLength > 20) {
+      return Response.exceptionHandling(342, "mission name should be between 1 and 20 characters");
+    }
+
     BeanUtils.copyProperties(missionInfo, mission.get());
     missionRepository.save(mission.get());
     return Response.ok();
@@ -120,11 +121,12 @@ public class MissionController {
 
 
   @GetMapping("/get/{missionId}")
-  @Operation(summary = "查看mission详情")
+  @Operation(summary = "get mission information")
   public Response<?> getMission(@PathVariable Long missionId) {
     Optional<Mission> mission = missionRepository.findById(missionId);
-//  UserSelfVO userSelfVO = new UserSelfVO();
-//  BeanUtils.copyProperties(user.get(), userSelfVO);
+    if (mission.equals(Optional.empty())) {
+      return Response.exceptionHandling(301, "mission does not exist");
+    }
     return Response.ok(mission);
   }
